+++ workspaces/cli-shared/src/streams/index.ts
@@ -1,1 +1,2 @@
 export * as DiffResults from './diff-results';
+export * as UndocumentedUrls from './undocumented-urls';
